# backend/userservice/api/api.yaml
openapi: 3.0.0
info:
    version: 1.0.0
    title: User Service
    description: User Service API

servers:
    - url: http://localhost:3001/

paths:
    /users:
        get:
            summary: Get all users
            responses:
                "200":
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/User"
        post:
            summary: Create user
            parameters:
                - in: header
                  name: internal-service
                  schema:
                      type: string
                  required: true
            requestBody:
                description: User to create
                required: true
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/User"
            responses:
                "201":
                    description: A list of users
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "403":
                    description: Not authorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string

    /users/{id}:
        get:
            summary: Get a user by ID
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: integer
            responses:
                "200":
                    description: A user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "404":
                    description: User not found
    /users/byDocId/{id}:
        get:
            summary: Get a user by Document ID
            parameters:
                - in: path
                  name: id
                  required: true
                  schema:
                      type: string
                - in: header
                  name: internal-service
                  required: true
                  schema:
                      type: string
            responses:
                "200":
                    description: A user
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
                "404":
                    description: User not found
                "403":
                    description: Not authorized
                    content:
                        application/json:
                            schema:
                                type: object
                                properties:
                                    message:
                                        type: string

components:
    schemas:
        User:
            type: object
            required:
                - steamId
                - name
                - avatar
                - profileUrl
                - steamMemberSince
            properties:
                steamId:
                    type: string
                name:
                    type: string
                avatar:
                    type: string
                profileUrl:
                    type: string
                steamMemberSince:
                    type: integer
        CreatedUser:
            allOf:
                - type: object
                  properties:
                      id:
                          type: string
                - $ref: "#/components/schemas/User"